renamed components + changed exports/importd
