+++ frontend/src/components/HeaderTitle/HeaderTitle.jsx
@@ -1,9 +1,9 @@
 import React from "react";
 
-const Header = ({ title }) => {
+const HeaderTitle = ({ title }) => {
   return (
     <h1 className="text-2xl xl:text-3xl font-extrabold text-black">{title}</h1>
   );
 };
 
-export default Header;
+export default HeaderTitle;
